Pin CI lint versions and make the silent gates actually verify
CI was red for 5 runs straight, but not because of our code. The lint
job ran `uvx ruff` (unpinned = always latest) while local was 0.15.22,
and ruff 0.16.0 widened its default rule set from 118 rules (E+F) to
413. A commit-free repo could go red on its own. Local-green/CI-red is
itself the defect: a local gate that cannot predict CI is not a gate.

- Pin RUFF_VERSION and PYRIGHT_VERSION in the workflow; bumping them is
  now a reviewable commit rather than an unannounced upstream event.
  Enumerating 413 rules was the alternative — unmaintainable, and it
  would forfeit upstream improvements.
- Add a root ruff.toml: unifies target-version (only chart-history and
  fandom-pulse had a pyproject.toml, so only those two inferred py311)
  and stops config discovery inside the repo, so a developer's global
  ruff config can no longer change results.
- Fix all 42 findings. Stale `# noqa: S310` directives are dropped but
  their trust rationale is kept as a plain comment. S310 stays off: it
  flags `Request(...)` constructors too, where nothing is opened yet.

Two gates were passing while verifying nothing:

- The secret scan died on checkout's shallow clone and reported
  `scanned ~0 bytes` — green. Fixed with fetch-depth: 0. The earlier
  claim that there were no leaks rested on that empty scan; a pattern
  sweep over all 17 commit trees now backs it up, but gitleaks' full
  rule set runs for real for the first time here.
- snapshot-gate and schema-validate now fail when they match zero
  files. A green gate that checked nothing is worse than a red one.

Reviving `needs: python` also un-skipped two jobs that had not run in
8 days, and both failed for real reasons:

- entity.schema.json forbade additional properties while entities.json
  carried debut/agency/wd_id — first-class fields produced by
  entities.py and consumed by report.py's rookie/agency rollups.
- The snapshot gate filtered on `provenance` alone, so signal-series
  documents were being validated against the snapshot schema. The
  filter now matches snapshot-schema's required keys exactly.

signal-series still has no JSON schema package, so its four fixtures
remain unverified by any contract. Left as separate work (D-030).

+++ modules/sonic-profile/src/sonic_profile/features.py
@@ -39,8 +39,8 @@ def _pulse_clarity(onset_env: np.ndarray, sr: int, hop: int) -> float:
         return 0.0
     ac = ac / ac[0]
     # 30~300 BPM에 해당하는 lag만 본다(그 밖의 주기는 박이 아님)
-    lo = max(1, int(round(60.0 / 300.0 * sr / hop)))
-    hi = min(ac.size - 1, int(round(60.0 / 30.0 * sr / hop)))
+    lo = max(1, round(60.0 / 300.0 * sr / hop))
+    hi = min(ac.size - 1, round(60.0 / 30.0 * sr / hop))
     if hi <= lo:
         return 0.0
     return float(np.clip(np.max(ac[lo : hi + 1]), 0.0, 1.0))
